ENH: use sqlite3 as the backing for the library instead of dbm
sqlite3, like dbm is built into the stdlib. Unlike dbm, sqlite3 files are
extremely portable. This means that a library generated on one machine may be
used on any other machine that slider supports. This can dramatically speed up
server deploys by allowing us to build a library once on a CI machine like
jenkins or travis and use that artifact in our deploys.

To assist with making libraries portable, paths are now stored relative to the
'.slider.db' file. This means that the entire library directory could be
relocated without needing to rebuild or adjust any values. The relocatability
allows us to create a tar archive of a library and ship this around to any other
servers regardless of where it will be unpacked.