TST: break up "mega tests" in `test_beatmap`
"Mega tests" occur when a single test is testing multiple independent things. In
general, a single test should test a single thing. This makes it much easier to
diagnose a problem by looking at the set of test failures. Imagine making a
change and seeing: `test failed: test_everything`. This doesn't tell the
programmer _what_ is wrong, just that something is wrong. Instead, it might be
better to see: `test failed: test_beatmap.test_ad`, this tells the programmer
that something is likely wrong with the `Beatmap.ad` method. Maybe something is
wrong with the beatmap parser getting the ad value, but at least it has been
narrowed down to something being wrong with ad. The other test failures should
help us narrow in on the actual root cause.

I realize there aren't enough tests in slider to make this really work yet;
however, it is better to be in the habit of writing small tests so that we don't
need to break them up later.

This PR breaks up the two tests in `test_beatmap.py` into one test per feature.
I am currently leaving each beatmap section (Metadata, HitObjects, General,
etc.) as a single test because it is unlikely that one thing in a section could
be broken without all of them being broken.

This PR also introduces a `pytest.fixture` to create the beatmap being examined.
`pytest.fixture` allows one to reuse test set up so that each of these little
tests doesn't need to start with the same beatmap parsing code.