MAINT: use example data machinery in tests
Slider comes with some example data, like beatmap files. These were mainly to
make it easy to write documentation that would work without requiring the user
to download some beatmaps and put them in the correct place. Instead, the docs
can reference `slider.example_data.*` and it will work without issues.

This commit moves the data that was in `tests/data` into the example data and
makes the test use this machinery. This standardizes the way that slider
incorporates example data into the project.

Right now, I added all of the difficulties of the particular MIIRO vs. Ai no
Scenario beatmap that was used in the existing tests. There are a lot of
difficulties for this map. We may want to change the tests to be programmed
against the existing example beatmap and remove this. I generally find
refactoring tests to be dangerous, and I personally don't think the extra space
used by these beatmaps is an issue, but I am not sure what other users of slider
think.